% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessFitMatrixPLM.R
\name{goodnessFitMatrixPLM}
\alias{goodnessFitMatrixPLM}
\title{Table of goodness of fit measures for (panel) linear models models.}
\usage{
goodnessFitMatrixPLM(x, decim = 4, decim_per = 2)
}
\arguments{
\item{x}{An object of class lm, plm.}

\item{decim}{Specifies the number of decimals to display. The default is 4.}

\item{decim_per}{Specifies the number of decimals to display in case of percent results. The default is 2.}
}
\value{
An n x 1 matrix with the R-squared, Adjusted R-squared, and F-statistic (p-value).
Note that for Fixed-Effects Models, withih R-squareds are provided as given by the plm().
}
\description{
Table of goodness of fit measures for (panel) linear models models.
}
\examples{
library(plm)
data("Grunfeld", package = "plm") ###use the models from plm package (Croissant Y. et al, 2022: pg.4):
pooledOLS=plm(inv ~ value + capital, data = Grunfeld, model="pooling")
goodnessFitMatrixPLM(pooledOLS)
FEM=plm(inv ~ value + capital, data = Grunfeld, model="within", effect="twoways")
goodnessFitMatrixPLM(FEM, 2, 2)
DVOLS=plm(inv ~ value + capital + factor(firm) + factor(year), data = Grunfeld, model="pooling")
goodnessFitMatrixPLM(DVOLS, 2, 2)
}
\references{
Croissant, Y. et al.(2022), Package ‘plm’:linear models for panel data (Version 2.6-2).https://cran.r-project.org/package=plm
}
